requirements.txt
# Core ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0

# Feature Engineering
category-encoders==2.6.1
statsmodels==0.14.0

# API Framework
fastapi==0.100.0
uvicorn==0.23.2
pydantic==1.10.12
python-multipart==0.0.6

# Orchestration
prefect==2.11.5

# Testing
pytest==7.4.2
pytest-cov==4.1.0

# Monitoring (simplified)
evidently==0.4.10

# MLOps
mlflow==2.7.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.8.5

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Logging
loguru==0.7.2

# Additional
python-docx==1.1.0
openpyxl==3.1.2
joblib==1.3.2